Add worker, fix postgres volumes

+++ app/manage.py
@@ -4,11 +4,20 @@ from app import app, db, Check
 
 cli = FlaskGroup(app)
 
-@cli.command('create_db')
-def create_db():
-	db.drop_all()
-	db.create_all()
-	db.session.commit()
+# @cli.command('create_db')
+# def create_db():
+# 	db.drop_all()
+# 	db.create_all()
+# 	db.session.commit()
+
+@cli.command('init_db')
+def init_db():
+	res = db.engine.dialect.has_table(db.engine, 'checks')
+	if not res:
+		print('Creating {} table'.format('checks'))
+		db.drop_all()
+		db.create_all()
+		db.session.commit()
 
 @cli.command('seed_db')	
 def seed_db():
